Replace shell script version of git-firkdiff with a Git alias.
Now forkdiff accepts the two branch names as first two parameters,
instead of last two. This enables us to pass any arbitrary git-diff
parameters to the command as is.

This change allowed me to get rid of the $# usage in the shell script,
which in turn allowed me to put it in a Git alias and not need an
external shell script.