Additions to $HOME/.bashrc
Added a single letter alias for Docker; Docker is here to stay, so make
life easy for my fingers :)

Add an alias to see only the tail of the `ll` command. I found myself
doing this very often to see the latest files that were changed: `ll |
tail`.

Make ping_google not try to resolve hostname; a bit of efficieny, and
almost consistent output (rather than longer/shorter lones depending on
how deep in the internet the target server is).

Make vi a function so that Ctrl-S command inside Vim does not cause the
shell to interfere.