Conditionally disable stdout and stderr in .bashrc
If the session is non-interactive, suppress the stdout and stderr of the
commands executed in ~.bashrc. This is to prevent the non-interactive
commands from getting confused by this output.

Specifically, scp command, which uses stdout for file transfer, was
confused by the output of the grep command.

In passing, also prevent the output of grep from landing on our screens.