Fix another bug in ll, introduced in previous commit
A plain `ll` command errors out with

  ls: cannot access '': No such file or directory

The lesson is that the folowing two snippets of code are not equivalent,
when the function is called without any parameters.

  function f() { local x="$@"; ls "$x"; }
  function f() { ls "$@"; }

When the function is called without any parameters, the first variant
still passes a single parameter, an empty string, to `ls`. Whereas the
second variant invokes `ls` without any parameters.