Do NOT source system/global bashrc
In retrospect, this was a horrible thing to do. The global file is
supposed to source the per-user .bashrc, not the other way around.
Sourcing the system's bashrc caused recursion, because it then sources
the user's .bashrc. But I had been saved so far from this infinite
recursion by a protection implemented in the system bashrc, where it
prevents itself from being sourced twice. Perhaps that protection was
implemented to protect the system from idiots like myself ;-)

This has been present since the first commit, so I don't have any
history or recollection of why I added this command in the first place.