Disabled the blue dot on pinned tabs in Firefox
When the title of a pinned tab changes, Firefox shows a blue dot on the
tab. In my case, it occurs most often for my GMail pinned tab.

The configuration disbles the blue-dot marker for all websites, not just
GMail. The answer linked to in comments has a fix, to target it to just
GMail, but since I don't want any of my pinned tabs to show the
blue-dot, the current code works as intended.

Added userChrome.css to Firefox profile; the profile path changes from
machine to machine, so this will need to be changed for the next machine
I'll need to use it on.