In tmux, use Ctrl+l shortcut to swith panes
This shortcut switches the panes while retaining window's zoom level.
This replaces the Alt/Meta+; shortcut, which stopped working for me
becuase macOS captures the Option/Alt+; and sends ellipses character to
the application. I tried hard to find a way to disable this behaviour in
macOS, but nothing worked, so I chose to simply use a different key
combination. Oddly, I believe I was using macOS back when I initially
added this shortcut.

This change also removes the cryptic tmux commands with a built-in tmux
command. I presume this built-in command was not available back when I
added the Alt/Meta+; shortcut.